+++ crates/zng-ext-input/src/drag_drop.rs
@@ -611,18 +611,6 @@ event! {
 }
 
 impl DropArgs {
-    /// Deprecated
-    #[deprecated = "use `target.contains_enabled`"]
-    pub fn is_enabled(&self, widget_id: WidgetId) -> bool {
-        self.target.contains_enabled(widget_id)
-    }
-
-    /// Deprecated
-    #[deprecated = "use `target.contains_disabled`"]
-    pub fn is_disabled(&self, widget_id: WidgetId) -> bool {
-        self.target.contains_disabled(widget_id)
-    }
-
     /// Stop propagation and set the `effect` that was applied to the data.
     ///
     /// Logs an error if propagation is already stopped.